Previous import state management
Adds a module to track the state of a slack to discord import.

This allows the program to keep track of the last message synced to discord.
Without this, a multi-channel import that fails partway through (which is quite common
for very large slacks) needs to have the export manually modified. Start timestamps will ignore
earlier messages for all channels, but the import only handles one channel at a time.

This allows idempotently re-running the slack-to-discord without worrying about
sending duplicate messages.

It is also necessary groundwork to allowing slack-to-discord to simultaneously import
all channels in the workspace, instead of attempting one at a time.