Refactor to run each channel concurrently
This takes the suggestion from this gh issue: https://github.com/pR0Ps/slack-to-discord/issues/35

And runs with it as-is. _import_channel includes the exact code that used to exist in the per-channel for loop block as its own async function, and all channels are run in tandem as a task group.

This PR also builds on top of the previous commit that adds simple json file-based state management, allowing idempotent re-runs. State management is a blocker for this task due to the unreliable speed for each channel -- a failure would leave each channel imported up to potentially different timestamps, rendering the start date parameter useless.